Fixed output issue with DescentHelper checking Int MAX instead of Double MAX

+++ src/main/java/games/descent2e/DescentHelper.java
@@ -694,7 +694,7 @@ public class DescentHelper {
                 minDistance = distance;
             }
         }
-        if (minDistance == Integer.MAX_VALUE) return -1;
+        if (minDistance == Double.MAX_VALUE) return -1;
         return minDistance;
     }
